src/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
httpx==0.28.1
loguru==0.7.3
pyjwt==2.10.1
fastapi-sessions==0.3.2